migrate private_ip_mode
- Removed `private_ip_mode` from `container` classes
- Created new `private_ip_mode` class and test
