+++ mongodbatlas/containers.go
@@ -12,20 +12,9 @@ type ContainerService struct {
 	sling *sling.Sling
 }
 
-// PrivateIPModeService provides many needfuls
-type PrivateIPModeService struct {
-	sling *sling.Sling
-}
-
 // newContainerService returns a new ContainerService.
 func newContainerService(sling *sling.Sling) *ContainerService {
 	return &ContainerService{
-		sling: sling.Path("groups/"),
-	}
-}
-
-func newPrivateIPModeService(sling *sling.Sling) *PrivateIPModeService {
-	return &PrivateIPModeService{
 		sling: sling.Path("groups/"),
 	}
 }
@@ -40,11 +29,6 @@ type Container struct {
 	GcpProjectID   string `json:"gcpProjectId,omitempty"`
 	NetworkName    string `json:"networkName,omitempty"`
 	Provisioned    bool   `json:"provisioned,omitempty"`
-}
-
-// PrivateIPMode is the response from both Enable/DisablePrivateIPMode
-type PrivateIPMode struct {
-	Enabled bool `json:"enabled,omitempty"`
 }
 
 // containerListResponse is the response from the ContainerService.List.
@@ -99,31 +83,5 @@ func (c *ContainerService) Delete(gid string, id string) (*http.Response, error)
 	apiError := new(APIError)
 	path := fmt.Sprintf("%s/containers/%s", gid, id)
 	resp, err := c.sling.New().Delete(path).Receive(container, apiError)
-	return resp, relevantError(err, *apiError)
-}
-
-// EnablePrivateIPMode enables needfuls
-// https://docs.atlas.mongodb.com/reference/api/set-private-ip-mode-for-project/
-func (p *PrivateIPModeService) EnablePrivateIPMode(gid string) (*http.Response, error) {
-	privateIPMode := new(PrivateIPMode)
-	apiError := new(APIError)
-	path := fmt.Sprintf("%s/privateIpMode", gid)
-	params := PrivateIPMode{
-		Enabled: true,
-	}
-	resp, err := p.sling.New().Patch(path).BodyJSON(params).Receive(privateIPMode, apiError)
-	return resp, relevantError(err, *apiError)
-}
-
-// DisablePrivateIPMode enables needfuls
-// https://docs.atlas.mongodb.com/reference/api/set-private-ip-mode-for-project/
-func (p *PrivateIPModeService) DisablePrivateIPMode(gid string) (*http.Response, error) {
-	privateIPMode := new(PrivateIPMode)
-	apiError := new(APIError)
-	path := fmt.Sprintf("%s/privateIpMode", gid)
-	params := PrivateIPMode{
-		Enabled: false,
-	}
-	resp, err := p.sling.New().Patch(path).BodyJSON(params).Receive(privateIPMode, apiError)
 	return resp, relevantError(err, *apiError)
 }
